refactor(workout-spa-editor): Extract format target helpers into separate modules
- Extract formatPowerTarget and formatHeartRateTarget into dedicated format-power-hr.ts module
- Extract formatCadenceTarget and formatPaceTarget into dedicated format-cadence-pace.ts module
- Create shared format-helpers.ts with isValidNumber utility function
- Update format-target-helpers.ts to re-export functions for backward compatibility
- Add comprehensive type checking with isValidNumber validation for all numeric values
- Improve code organization and maintainability by separating concerns into focused modules
