+++ packages/workout-spa-editor/src/components/molecules/StepCard/format-helpers.ts
@@ -1,0 +1,6 @@
+/**
+ * Check if a value is a valid finite number
+ */
+export const isValidNumber = (value: unknown): value is number => {
+  return typeof value === "number" && Number.isFinite(value);
+};
